Removing logging of `User#cas_extra_attributes=`
Based on testing against test CAS, there are no additional attributes
provided. I'm leaving the method as a placeholder and reminder that
there could be information in the future.

It would be great to capture additional information from CAS and assign
that to the User record that is created. This is based on the
[`devise_cas_authenticatable` gem's documentation][documentation]

[documentation]:https://github.com/nbudin/devise_cas_authenticatable#extra-attributes